requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==3.1.2
python-decouple==3.6